moved SharedPreferenceDataSource to local-source module
