+++ app/src/testShared/java/com/freshdigitable/udonroad2/TestAppComponent.kt
@@ -19,8 +19,8 @@ package com.freshdigitable.udonroad2
 import android.app.Application
 import com.freshdigitable.udonroad2.data.db.DatabaseModule
 import com.freshdigitable.udonroad2.data.db.dao.UserDao
-import com.freshdigitable.udonroad2.data.impl.SharedPreferenceDataSource
 import com.freshdigitable.udonroad2.data.impl.di.RepositoryModule
+import com.freshdigitable.udonroad2.data.local.SharedPreferenceDataSource
 import com.freshdigitable.udonroad2.data.restclient.AppTwitterModule
 import com.freshdigitable.udonroad2.di.ActivityBuilders
 import com.freshdigitable.udonroad2.di.AppComponent
