+++ data/repository/src/main/java/com/freshdigitable/udonroad2/data/impl/OAuthTokenRepository.kt
@@ -16,6 +16,7 @@
 
 package com.freshdigitable.udonroad2.data.impl
 
+import com.freshdigitable.udonroad2.data.local.SharedPreferenceDataSource
 import com.freshdigitable.udonroad2.data.restclient.OAuthApiClient
 import com.freshdigitable.udonroad2.model.AccessTokenEntity
 import com.freshdigitable.udonroad2.model.RequestTokenItem
