+++ data/repository/src/main/java/com/freshdigitable/udonroad2/data/impl/TweetRepository.kt
@@ -2,6 +2,7 @@ package com.freshdigitable.udonroad2.data.impl
 
 import com.freshdigitable.udonroad2.data.db.DaoModule
 import com.freshdigitable.udonroad2.data.db.dao.TweetDao
+import com.freshdigitable.udonroad2.data.local.SharedPreferenceDataSource
 import com.freshdigitable.udonroad2.data.restclient.TweetApiClient
 import com.freshdigitable.udonroad2.model.TweetId
 import com.freshdigitable.udonroad2.model.app.AppTwitterException
